seccion de reportes mejorada
