+++ src/components/admin/report/types/pdfTypes.ts
@@ -1,0 +1,17 @@
+import { AutoTableSettings } from "@/types/admin";
+
+
+declare module 'jspdf' {
+    interface jsPDF {
+        lastAutoTable: {
+            finalY: number;
+        };
+        autoTable: (options: AutoTableSettings) => jsPDF;
+    }
+}
+
+export interface PdfGeneratorOptions {
+    fontSize?: number;
+    theme?: string;
+    columnStyles?: Record<string, unknown>;
+}
